+++ CHANGES.md
@@ -32,6 +32,7 @@ Changes in CUPS v2.3b8
 - Localization updates (Issue #5461, Issues #5471, Issue #5481, Issue #5486,
   Issue #5489, Issue #5491, Issue #5492, Issue #5493, Issue #5494, Issue #5495,
   Issue #5497, Issue #5499, Issue #5500, Issue #5501, Issue #5504)
+- The scheduler did not always idle exit as quickly as it could.
 
 
 Changes in CUPS v2.3b7